# server/api/schemas.py
"""Request and response models (these also define the OpenAPI schema the frontend uses)."""

from __future__ import annotations

from typing import Literal

from pydantic import BaseModel, Field, model_validator

from server import engines
from server.engines.base import BLEND_RANGE, PITCH_RANGE, SPEED_RANGE


class ErrorBody(BaseModel):
    code: Literal[
        "invalid_input",
        "too_long",
        "quota_exceeded",
        "rate_limited",
        "unauthorized",
        "busy",
        "not_found",
        "internal",
    ]
    message: str


class ErrorResponse(BaseModel):
    error: ErrorBody


# ── Voices ───────────────────────────────────────────────────────────────────
class VoiceOut(BaseModel):
    id: str
    name: str
    gender: Literal["female", "male"]
    grade: str | None
    tags: list[str] = Field(default_factory=list, description='e.g. "native"')


class LanguageOut(BaseModel):
    code: str
    name: str
    engine: Literal["kokoro", "indic_mio"]
    default_voice: str
    preview_text: str
    voices: list[VoiceOut]
    emotions: list[str] = Field(
        default_factory=list, description="Emotion tags this language understands, e.g. happy"
    )


class VoicesResponse(BaseModel):
    languages: list[LanguageOut]
    speed: tuple[float, float] = SPEED_RANGE
    pitch: tuple[float, float] = PITCH_RANGE


# ── Jobs ─────────────────────────────────────────────────────────────────────
class Pronunciation(BaseModel):
    word: str = Field(min_length=1, max_length=100)
    say: str = Field(min_length=1, max_length=200)


class JobCreate(BaseModel):
    text: str = Field(min_length=1, description="Text to speak (limit depends on the account)")
    lang: str = Field("a", description="Language code, see /v1/voices")
    voice: str = "af_heart"
    blend_voice: str | None = None
    blend_ratio: float = Field(0.5, ge=BLEND_RANGE[0], le=BLEND_RANGE[1])
    speed: float = Field(1.0, ge=SPEED_RANGE[0], le=SPEED_RANGE[1])
    pitch: float = Field(0.0, ge=PITCH_RANGE[0], le=PITCH_RANGE[1])
    pronunciations: list[Pronunciation] = Field(default_factory=list, max_length=200)
    turnstile_token: str | None = Field(
        None, max_length=4096, description="Required for anonymous visitors"
    )

    @model_validator(mode="after")
    def _check_voice(self) -> JobCreate:
        if not self.text.strip():
            raise ValueError("There is no text to speak.")
        engines.validate(self.lang, self.voice, self.blend_voice)  # ValueError → 422
        return self


class JobOut(BaseModel):
    id: str
    status: Literal["queued", "running", "done", "error", "cancelled"]
    progress: int = 0
    queue_position: int | None = None
    estimated_seconds: float | None = None
    audio_seconds: float | None = None
    url: str | None = Field(None, description="Signed MP3 link, valid for an hour (when done)")
    wav_url: str | None = Field(None, description="Signed WAV link (signed-in users)")
    error: str | None = None
    stream_url: str


# ── Text ─────────────────────────────────────────────────────────────────────
class TextIn(BaseModel):
    text: str = Field(max_length=2_000_000)


class TextOut(BaseModel):
    text: str


class ExtractOut(BaseModel):
    text: str
    cleaned: bool
    characters: int


class LimitsOut(BaseModel):
    max_chars: int
    daily_chars: int
    max_active_jobs: int
    uploads: bool
    wav: bool


class MeOut(BaseModel):
    signed_in: bool
    email: str | None
    limits: LimitsOut
    chars_today: int


class HistoryItem(BaseModel):
    id: str
    created_at: str
    lang: str
    voice: str
    blend_voice: str | None
    blend_ratio: float | None
    speed: float
    pitch: float
    chars: int
    audio_seconds: float | None
    url: str | None
    wav_url: str | None


class HistoryOut(BaseModel):
    items: list[HistoryItem]
    days: int


class PronunciationsBody(BaseModel):
    entries: list[Pronunciation] = Field(max_length=200)

    @model_validator(mode="after")
    def _unique_words(self) -> PronunciationsBody:
        words = [e.word.strip().lower() for e in self.entries]
        if len(words) != len(set(words)):
            raise ValueError("Each word can only have one pronunciation.")
        return self


class ConfigOut(BaseModel):
    max_chars: int = Field(description="Characters per request for this visitor")
    max_upload_mb: int
    document_types: list[str]


class HealthOut(BaseModel):
    status: Literal["ok", "degraded"]
    redis: Literal["ok", "down"]
    workers: int