# web/src/api/openapi.json
{
  "components": {
    "schemas": {
      "Body_extract_v1_files_extract_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_extract_v1_files_extract_post",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "enum": [
              "invalid_input",
              "too_long",
              "quota_exceeded",
              "rate_limited",
              "unauthorized",
              "busy",
              "not_found",
              "internal"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExtractOut": {
        "properties": {
          "characters": {
            "title": "Characters",
            "type": "integer"
          },
          "cleaned": {
            "title": "Cleaned",
            "type": "boolean"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "cleaned",
          "characters"
        ],
        "title": "ExtractOut",
        "type": "object"
      },
...
      "TextIn": {
        "properties": {
          "text": {
            "maxLength": 2000000,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TextIn",
        "type": "object"
      },
      "TextOut": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TextOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Text-to-speech jobs with live streaming. Errors always look like `{\"error\": {\"code\": ..., \"message\": ...}}`.",
    "title": "Narravo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/files/extract": {
      "post": {
        "description": "Text from a .txt, .md, .docx or .pdf file (up to 5 MB; signed-in users).\n\nPDF text comes back cleaned.",
        "operationId": "extract_v1_files_extract_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_v1_files_extract_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Extract",
        "tags": [
          "text"
        ]
      }
    },
...
    "/v1/text/clean": {
      "post": {
        "description": "Fix pasted text: quotes, links, markdown symbols, lines broken mid-sentence.",
        "operationId": "clean_v1_text_clean_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Clean",
        "tags": [
          "text"
        ]
      }
    },
...
  }
}